+++ src/modules/overrepresented_seqs.rs
@@ -322,6 +322,10 @@ impl OverRepresentedSeqs {
 }
 
 impl QCModule for OverRepresentedSeqs {
+    fn cost_hint(&self) -> u32 {
+        6
+    }
+
     fn process_sequence(&mut self, sequence: &Sequence) {
         self.computed = None;