+++ src/modules/per_base_quality.rs
@@ -165,6 +165,10 @@ impl PerBaseQualityScores {
 }
 
 impl QCModule for PerBaseQualityScores {
+    fn cost_hint(&self) -> u32 {
+        8
+    }
+
     fn process_sequence(&mut self, sequence: &Sequence) {
         let qual = &sequence.quality;